Use `setup_test_component_platform` helper for update entity component tests instead of `hass.components`
